requirements.txt
Flask==2.0.1
cassandra-driver==3.27.0
gunicorn==20.1.0